renamed graph sandbox to sandbox
