+++ client/index.html
@@ -2,11 +2,26 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" href="/favicon.ico" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta property="og:title" content="Magic CS" />
-    <meta property="og:description" content="Visualize with Magic Algorithms" />
-    <meta property="og:type" content="website" />
+    <link
+      rel="icon"
+      href="/favicon.ico"
+    />
+    <meta
+      name="viewport"
+      content="width=device-width, initial-scale=1.0"
+    />
+    <meta
+      property="og:title"
+      content="Magic CS"
+    />
+    <meta
+      property="og:description"
+      content="Visualize with Magic Graphs"
+    />
+    <meta
+      property="og:type"
+      content="website"
+    />
     <meta
       property="og:image"
       content="public/products/thumbnails/graph-playground.png"
@@ -15,16 +30,25 @@
       property="twitter:image"
       content="public/products/thumbnails/graph-playground.png"
     />
-    <meta property="twitter:card" content="summary" />
-    <meta property="twitter:title" content="Magic CS" />
+    <meta
+      property="twitter:card"
+      content="summary"
+    />
+    <meta
+      property="twitter:title"
+      content="Magic CS"
+    />
     <meta
       property="twitter:description"
-      content="Visualize with Magic Algorithms"
+      content="Visualize with Magic Graphs"
     />
-    <title>Magic CS ✨🦄</title>
+    <title>Magic Graphs ✨🦄</title>
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="/src/main.ts"></script>
+    <script
+      type="module"
+      src="/src/main.ts"
+    ></script>
   </body>
 </html>
