+++ client/src/products/sandbox/ui/AnnotationToolbar.vue
