+++ client/src/products/sandbox/ui/CollaborativeSessionMenu.vue
