+++ client/src/products/sandbox/ui/GraphInfoMenu/AdjacencyList.vue
