+++ client/src/products/sandbox/ui/GraphInfoMenu/AdjacencyListDisplay.vue
