+++ client/src/products/sandbox/ui/GraphInfoMenu/ConnectedInfo.vue
