+++ client/src/products/sandbox/ui/GraphInfoMenu/GraphInfoMenu.vue
