+++ client/src/products/sandbox/ui/GraphInfoMenu/TransitionMatrixLabel.vue
