+++ client/src/products/sandbox/ui/GraphInfoMenu/useBipartiteColorizer.ts
