+++ client/src/products/sandbox/ui/GraphInfoMenu/useCycleColorizer.ts
