+++ client/src/products/sandbox/ui/GraphInfoMenu/useSCCColorizer.ts
