+++ client/src/products/sandbox/ui/MarkupMenu/InputColor.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-  import { MARKUP_COLORS } from '@product/graph-sandbox/constants';
+  import { MARKUP_COLORS } from '@product/sandbox/constants';
   import type { Color } from '@utils/colors';
   import ButtonGroup from './ButtonGroup.vue';
   import GRadioButton from '@ui/graph/button/GRadioButton.vue';
