+++ client/src/products/sandbox/ui/MarkupMenu/InputSize.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
-  import { MARKUP_SIZES } from '@product/graph-sandbox/constants';
-  import type { MarkupSize } from '@product/graph-sandbox/types';
+  import { MARKUP_SIZES } from '@product/sandbox/constants';
+  import type { MarkupSize } from '@product/sandbox/types';
   import ButtonGroup from './ButtonGroup.vue';
   import GRadioButton from '@ui/graph/button/GRadioButton.vue';
